pyproject.toml
[project]
name = "sopp"
dynamic = ["version"]
dependencies = [
    "sgp4~=2.21",
    "skyfield>=1.53",
    "requests~=2.32.4",
    "python-dateutil~=2.8.2",
    "python-dotenv~=0.21.0",
    "pyarrow>=14.0.0",
    "pandas>=2.0.0",
]
requires-python = ">=3.8"
description = "SOPP is an open-source tool for calculating satellite interference to radio astronomy observations."
readme = "README.md"
license = "AGPL-3.0-or-later"
keywords = ["satellites", "astronomy"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[project.optional-dependencies]
cli = [
    "rich>=13.0.0",
    "typer>=0.9.0",
]
docs = [
    "mkdocs-material",
    "mkdocstrings[python]",
    "ruff",
]

[project.urls]
Documentation = "https://nsf-swift.github.io/satellite-overhead"
Repository = "https://github.com/NSF-Swift/satellite-overhead"
"Bug Tracker" = "https://github.com/NSF-Swift/satellite-overhead/issues"

[project.scripts]
sopp = "sopp.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
path = "src"

[tool.hatch.version]
path = "src/sopp/__about__.py"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.envs.default]
python = "3.11"
features = ["cli"]

[tool.hatch.envs.test]
features = ["cli"]
dependencies = [
  "pytest~=7.2.0",
  "pytest-mock~=3.11.1",
]

[tool.hatch.envs.test.scripts]
run = "pytest {args:tests}"

[[tool.hatch.envs.test.matrix]]
python = ["3.10", "3.11", "3.12"]

[tool.hatch.envs.ci]
template = "test"
dev-mode = false

[tool.hatch.envs.docs]
features = ["docs"]
dependencies = []

[tool.hatch.envs.docs.scripts]
serve = "mkdocs serve"
build = "mkdocs build --strict"
deploy = "mkdocs gh-deploy --force"

[tool.hatch.envs.lint]
detached = true
dependencies = [
  "ruff",
]

[tool.hatch.envs.lint.scripts]
check = "ruff check . --ignore E501"
fmt = "ruff format --check --diff ."
all = ["check", "fmt"]

[tool.ruff]
target-version = "py310"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
